+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0-1.19.3]
+### Added
+- Add the ability to click through signs and item frames
+  - Sign click through is prevented when holding a sign, dye, an ink sac, or a glowing ink sac in your main hand
+  - Item Frame click through is disabled while sneaking
+
 ## [1.2.0-1.19.3]
 ### Changes
 - Update minecraft version to 1.19.3